Remove ancient codegen opt for expanded Scala anon function classes.
Since the very early days of Scala.js, we have had a code size
optimization for lambdas expanded as anonymous classes. We tried
very hard to recover the captures from fields, and the body of the
`apply` method, in order to reconstruct a `js.Closure`.

That made sense as long as we supported Scala 2.11 (and before
that, 2.10). However, since Scala 2.12.x, scalac almost never
expands lambdas of `scala.FunctionN`. Instead, they reach the
backend as `Function` node.

The only situations where scalac still expands lambdas as anonymous
classes is if it used in an argument to a super constructor (or
delegate `this()` constructor).

The code paths for that code size optimization are therefore almost
dead code, and do not pull their weight anymore. It was, AFAICT,
the only place where we still *tried* to emit something a certain
way, with a fallback.

This commit removes the optimization and all the related
infrastructure.