JUnit: populate sbt.testing.Event.throwable on test failure
If a Throwable associated with the test event is available, bubble it up through the SBT's test interface using the slot dedicated to just such information: `sbt.testing.Event.throwable`.

In addition to it being a shame to throw away information about the test failure that is literally in our hands, this simplifies integration with runners that assume that test failure is always accompanied by a `Throwable` (e.g., Gradle).