Fix #5131: Overhaul justifications for caching in Emitter.genClass.
Previously, we were missing a number of cases where we should
have set the `changed` flag to true.

That flag is used to decide if a module needs to be rewritten.
Previously, we failed to notice cases where *only* one of the
`Analysis` flags changed (`hasInstances`, `hasRuntimeTypeInfo`
and `hasInstanceTests`). In those cases, all the *cached trees*
are kept. The only thing that changes is the subset thereof that
is actullay included. That does not require invalidation of the
caches, but it does require rewriting the module.

Likewise, we failed to invalidate the module if only the *set*
of static-like methods retained by the analysis changed.

Overall, it had become quite tricky to follow what we are
allowed to use where. This commit overhauls how we track
everything.

First, we include the `kind` of the class in its cache identity,
because too many things depend on it. If it changes, it is
likely that everything will be invalidated anyway.

Other than that, dangerous variables, whose uses must be
justified, are clearly marked with `foo_!` names.

Boolean decisions whose outcome is never cached are tracked
separately; if any of them changes, we marked `changed` to `true`.
This particular change is what fixes #5131.

Finally, we add dedicated tracking of the set of generated
static-like methods, like we do for instance methods.