Fix #5131: Take Analysis flags into account for the `changed` flag.
That flag is used to decide if a module needs to be rewritten.
Previously, we failed to notice cases where *only* one of the
`Analysis` flags changed (`hasInstances`, `hasRuntimeTypeInfo` and
`hasInstanceTests`). In those cases, all the *cached trees* are
kept. The only thing that changes is the subset thereof that is
actullay included. That does not require invalidation of the
caches, but it does require rewriting the module.