Enforce that jl.Throwable must a class extending jl.Object.
The emitter relies on that to rewire `jl.Throwable` to extend JS
`Error`, and bridge methods to those of `jl.Object`.

It previously had a side condition that `jl.Throwable` actually
extends `jl.Object` to perform that rewrite. If that were false,
though, some of the guarantees we give about `Throwable`s would
be broken, so I'm pretty sure something would break somewhere.

It is safer to directly enforce that universal fact in the
`ClassDefChecker` in the first place.