Wasm: Implement fmod on the Wasm side.
This avoids the JS call overhead for that "primitive" operation.

As can be seen in the implementation, `fmod` is far from a
primitive, though. We took an MIT implemention of `fmod`, generic
in the bit-width, and translated it by hand to Wasm. We use the
newly introduced `GenericInstructions` to also have a generic
implementation. So we generate one function for `Float_%` and one
for `Double_%`.