Merge the itables into the corresponding vtables.
This is a trade-off on memory and code size.

The big advantage is that run-time instances of our classes use
one fewer word each, since we do not have an `itables` field
anymore. There could be some execution time improvements as a
side effect, since we have one fewer word to initialize.

There are two disadvantages, but IMO they are comparatively minor:

* We cannot use the "one empty itables to rule them all" trick
  anymore. Classes that implement no interfaces need to list their
  N `null` values. This has both a code size impact and a constant
  run-time memory cost.
* We cannot share the itables of array classes either. This has no
  impact on code size (there is a single place in the code where
  we initialize those fields), but it does add run-time footprint
  to the vtables of each array type that gets created.

Interface method calls are not impacted, neither in the code we
generate, nor in their run-time performance. Virtual method calls
are only impacted insofar as the index of virtual method pointers
is bumped up by the number of itable slots. This can add one more
byte to some calls, due to the var-length encoding. If we get very
unlucky, it could also push some vtable method pointers past a
pre-fetch boundary, slowing down the calls at run-time. But I
don't really think that's plausible.