Introduce a binary API with Scala functions in `Reflect`.
And make its internals independent of JS interop.
This way, it will be usable in Wasm without JS host.

We use new names for the methods, instead of overloads, because
older compilers look for those methods by name only. If an older
compiler gets a newer library on the classpath, and the methods
with those names became overloaded, the compiler would crash.

In this commit, we do not change the compiler yet, to test that the
deprecated methods are correct.