Use a DataView in FloatingPointBits.
Previously, we used several typed arrays pointing to the same
`ArrayBuffer`. Now, we use a single `DataView`. Since we can (must)
force a specific endianness with `DataView`, we get rid of the
`highOffset` and `lowOffset` fields, which further streamlines the
code.

The assembly produced by V8 is now better. Since it knows that it
manipulates a single buffer, it needs fewer loads and type tests
internally. I expect the same would be true for other optimizing
engines.