Introduce IR UnaryOps for floating point bit manipulation.
Previously, our floating point bit manipulation methods were
implemented in user space. This commit instead introduces dedicated
IR `UnaryOp`s for them.

On Wasm, the implementations are straightforward opcodes. In
JavaScript, we used the same strategies as before, but moved to the
linker world.

In most cases, we use a unique scratch `DataView` to perform the
manipulations. It is now allocated as a `globalVar` in the emitter,
rather than in user space. The functions for `Float`/`Int`
conversions are straightforward, as well as those for
`Double`/`Long` when we use `bigint`s for `Long`s. When we use
`RuntimeLong`s, the conversions are implemented in `RuntimeLong`,
but require the scratch `DataView` as an argument, which the
linker injects.

This new strategy brings several benefits:

* For JavaScript, the generated code is basically optimal now.
  It produces tight sequences of Assembly instructions that do not
  allocate anything.
* For Wasm, the implementation does not rely on JS interop anymore,
  even when the optimizer does not run. This property will be
  required when we target Wasm outside of a JS host.
* We open a path to adding support for the "raw" variants
  `floatToRawIntBits`/`doubleToRawLongBits` in Wasm in the future,
  should we need it.