Support reading jar files in the IR cleaner.
That allows the IR cleaner to see the JS type definitions from
the library when it appears as a jar on the classpath. This is
the case for the linker private library. Adding that support will
allow the linker private library to use JS type definitions from
the Scala.js library, as long as the references can be erased
away.