Refactoring: Isolate handling of javalib methods with special bodies.
A number of methods from the javalib are special-cased by the
compiler, which replaces their body with a dedicated `UnaryOp` or
`BinaryOp`.

This commit refactors that handling to isolate it better from the
handling of regular methods. We also make it a bit more flexible,
so that we can more easily add further such methods in the future.