Wasm: Implement PriorityQueue without js.Array in Wasm backend
Replaced the `js.Array`-based `PriorityQueue` backend with a `scala.Array` implementation when targeting Wasm.
The previous js.Array based PriorityQueue implementation was not very performant on Wasm, because of the overhead of JS-interop for each inner array operations.
A `linkTimeIf` condition now selects the appropriate array implementation at build time.

Benchmarks show a 2-4x speedup for `add`, `peek`, and `poll` operations on the Wasm backend.
https://github.com/tanishiking/scalajs-benchmarks/pull/3