Opt: Use unsigned arithmetics in Range, instead of Longs.
Previously, `Range` used a number of intermediate operations on
`Long`s to avoid overflow. The optimizer was already good enough
to remove them in many cases, in particular when the step is `1`
and/or the start is `0`, which are common cases.

Now that the optimizer can reason about unsigned division, we can
do a lot better. We can entirely avoid `Long` computations, *and*
streamline a lot of code, by using unsigned `Int` arithmetics.
This produces much better code for the cases where the optimizer
cannot entirely get rid of the computations.