Use static methods as entry points for RuntimeLong operator methods.
Previously, we used instance methods on `RuntimeLong` to implement
the operators whose lhs is a `RuntimeLong`. Other operators, such
as `fromDouble`, were implemented as methods of the module.

Now, we use static methods for all the operators. Once inline, it
does not make any difference. However, it leaves behind a few
unused static methods, instead of some instance methods. The former
can be removed by off-the-shelf JS minifies, unlike the latter.

Paradoxically, the `Minify` output, as we measure it, gets bigger.
That's because static method names are not renamed by the name
minifier.