Opt: Avoid the `Long` division in `RuntimeLong.toString`.
Through clever analysis of approximation errors, it turns out we
can rely on a `Double` division by 10^9 rather than a `Long`
division.

Since there was already a `Double` division (and remainder) at the
end of the `Long` division algorithm, the new implementation should
be strictly better. We do have a new call to `js.Math.floor`, but
that should be efficient, as `floor` is typically a hardware
instruction.

We also remove the hackish way of reusing `unsignedDivModHelper`
to compute `toString()`.

According to the `longMicro` benchmark, this change gives a 3x
speed improvement to this code path.