Introduce a Scala-agnostic API for Reflect.
The new API is offered as part of the `javalibintf`. It only
uses classes from the javalib, which makes it Scala-agnostic.

The old binary API used by the codegen calls the new (public)
API in javalibintf. We deprecate them, but keep them for
backward binary compatibility.

The public API of `scala.scalajs.reflect.Reflect` is left
untouched, and is still the recommended way to access the
public features from user-land Scala code.

In this commit, we do not change the compiler yet, to test that the
deprecated methods are correct.