Opt: Branchless addition, subtraction and negation for `RuntimeLong`.
Hacker's Delight offers branchless formulas for double-word
additions and subtraction, without access to the machine's carry
bit.

Although the new formula contains more elementary instructions,
removal of the branch is significant. When one operand is
constant, folding reduces to 3 bitwise operations to compute the
carry, which is very fast. When both operands are variable, then
the carry is often 50/50 unpredictable, which means the branch
is unpredictable, and removing it is worth the full 5 bitwise
operations anyway.

Negation does not need a special code path anymore. Regular
folding of the `0L - b` formula yields optimal, branchless code
anyway. We remove `RuntimeLong.neg` and its code paths in the
optimizer and emitter. While we're there, we also remove
`RuntimeLong.not`, since the regular code paths for `-1L ^ b`
fold in the same way.

This change reduces execution time of the `sha512` benchmark by a
whopping 25-30%.