Introduce IR ops for unsigned extension and comparisons.
This completes the set of `UnaryOp`s and `BinaryOp`s to directly
manipulate the unsigned representation of integers.

Unlike other operations, such as `Int_unsigned_/`, the unsigned
extension and comparisons have efficient (and convenient)
implementations in user land. It is common for regular code to
directly use the efficient implementation (e.g.,
`x.toLong & 0xffffffffL`) instead of the dedicated library method
(`Integer.toUnsignedLong`).

If we only replaced the body of the library methods with IR nodes,
we would miss improvements in all the other code.

Therefore, in this case, we instead recognize the user-space
patterns in the optimizer, and replace them with the unsigned IR
operations through folding. Moreover, for unsigned comparisons,
we also recognize the patterns in the compiler backend. The purpose
here is mostly to make sure that all these opcodes end up in the
serialized IR, so that we effectively test them along the entire
pipeline.

When targeting JavaScript, the new IR nodes do not actually make
any difference. For `int` operations, the Emitter sort of "undoes"
the folding of the optimizer to implement them. That said, it could
choose an alternative implementation based on `>>> 0`, which we
should investigate in the future. For `Long`s, the subexpressions
of the patterns are expanded into the `RuntimeLong` operations
before folding gets a chance to recognize them (when they have not
been transformed by the compiler backend). That's fine, because
internal folding of the underlying `int` operations will do the
best possible thing anyway.

The size increase is only due to the additional always-reachable
methods in `RuntimeLong`. Those can be removed by standard JS
minifiers.

When targeting Wasm, this allows the emitter to produce the
dedicated Wasm opcodes, which are more likely to be efficient.

To be fair, we could have achieved the same result by recognizing
the patterns in the Wasm emitter instead. The deeper reason to add
those IR operations is for completeness. They were the last
operations from a standard set that were missing in the IR.