Optimize divisions and remainders by constants.
Following the techniques described in Hacker's Delight, Chapter 10.

Benchmarking suggested that is this generally worth it, except for
non-powers of 2 for a) int operations on JS and b) long operations
on Wasm.

For the long divisions on JS, this yields a 6x speedup, despite the
fact that the resulting code needs 16 (!) primitive int
multiplications.