Use `x >>> 0` instead of `x ^ 0x80000000` for unsigned comparisons.
Benchmarks show that this is slightly faster. Inspection of the
source code of v8 also suggests that they do not recognize
`x ^ 0x80000000` as doing anything special, but they do recognize
`x >>> 0` as emitting an "`Unsigned32`", and they do generate
unsigned comparisons when both inputs are known to be `Unsigned32`.