Do not use `Double` arithmetics in `Integer.parseInt()`.
Only use `Int` arithmetics. To detect overflow, we precompute a
table of the maximum string length for each radix.

`Int` arithmetics are faster than `Double` arithmetics. The
previous code used `Double`s to have a concise way of detecting
the overflow, which is not bad on JS engines. Given a cached
overflow detection mechanism, resorting to `Double`s is not
necessary anymore.

We use a `scala.Array` to be Wasm-friendly, since everything else
in the algorithm is now otherwise Wasm-friendly.