Do not use `Double` arithmetics in `Integer.parseInt()`.
Only use `Int` arithmetics. To detect overflow, we compute an
overflow barrier in way that should typically be constant-folded.

`Int` arithmetics are faster than `Double` arithmetics. The
previous code used `Double`s to have a concise way of detecting
the overflow, which is not bad on JS engines. However, we some
careful analysis of the possible overflows, we can do better.

We split the implementation of `parseInt` and `parseUnsignedInt`,
since they have more differences than common parts at this point.
Moreover, the justifications are quite different in each.

The new algorithms are also much more Wasm-friendly.