Remove all user-space `asUint`/`toUint` in the libraries.
Instead, we use `Integer.toUnsignedLong(x).toDouble`, which is
semantically equivalent. The only remaining use of `x >>> 0` is in
the JS-only runtime libraries, notably `RuntimeLong`.

We add some optimizations to generate the best possible code on all
targets.

For JS with `RuntimeLong`, we need to mark `RuntimeLong.toDouble`
as fully inline. That is fine, as its body is actually very short
(shorter than most methods of `RuntimeLong`). Moreover, we need a
tailored rewrite in the optimizer to get rid of a `Double` addition
of `0.0 + y` when `y` is provably non-negative.

For JS with `bigint`s, we directly fold
`(double) <toLongUnsigned>(x)` into `x >>> 0`. Otherwise, we
unnecessarily go through a `bigint` with `Number(BigInt(x >>> 0))`.

For Wasm, we fold the same shape into a single operation
`f64.convert_i32_u`, which we add in `WasmTransients`.