Remove all user-space `asUint`/`toUint` in the libraries.
Instead, we use `Integer.toUnsignedLong(x).toDouble`, which is
semantically equivalent. The only remaining use of `x >>> 0` is in
the JS-only runtime libraries, notably `RuntimeLong`.

We add some optimizations to generate the best possible code on all
targets.

For JS with `RuntimeLong`, we need a tailored rewrite in the
optimizer to get rid of a `Double` addition of `0.0 + y` when `y`
is provably non-negative. This shape comes from inlining
`RuntimeLong.toDouble`.

For JS with `bigint`s, we directly fold
`(double) <toLongUnsigned>(x)` into `x >>> 0`. Otherwise, we
unnecessarily go through a `bigint` with `Number(BigInt(x >>> 0))`.

For Wasm, we fold the same shape into a single operation
`f64.convert_i32_u`, which we add in `WasmTransients`.

Overall, this has no real impact on the JS target. However, it
removes two round-trips from Wasm to JS. Previously, we needed to
call a helper function to do the `x >>> 0`, then call the `$uD`
helper to retrieve the result. Now, the same operations stays
entirely within Wasm.