Remove all user-space `asUint`/`toUint` in the libraries.
Instead, we use `Integer.toUnsignedLong(x).toDouble`, which is
semantically equivalent. The only remaining use of `x >>> 0` is in
the runtime libraries, notably `RuntimeLong`.

We need to mark `RuntimeLong.toDouble` as fully inline so that this
new idiom is not a regression. That is fine, as its body is
actually very short (shorter than most methods of `RuntimeLong`).

Moreover, we need a tailored rewrite in the optimizer to get rid of
a `Double` addition of `0.0 + y` when `y` is provably non-negative.