Mark `RuntimeLong.toDouble` as fully inline.
Its implementation is shorter than many of the methods we already
inline, like additions and multiplications. We do pay a little
code size cost for this change.

This will be particularly useful for the following commit, which
will rely on folding to happen within `toDouble`, once inlined at
call site.