Further optimize Math.rint using a floating point splitting technique.
Benchmarks show that this is markedly faster than the previous
implementation. That is unsurprising, as the `js.Math.round`
algorithm basically had to handle x.5 values *twice* on top of
the primitive hardware rounding function (once inside
`js.Math.round` to turn ties-to-even into ties-up; and once in
our algorithm to turn ties-up back to ties-to-even).