Redesign the encoding of Longs in JavaScript.
Overall, we make the flat representation of longs as a pair of
`(lo, hi)` *the* representation, at the Emitter level. There are no
more instances of `RuntimeLong`. The emitter flattens out all the
`Long`s as follows:

* A local variable of type `long` becomes two local variables of
  type `int`.
* A field of type `long` becomes two fields of type `int`.
* An `Array[Long]` is stored as an `Int32Array` with twice as many
  elements, alternating `lo` and `hi` words.
* Method parameters of type `long` are expanded as *two* parameters
  of type `int`.

For the *result* of method parameters, there are several competing
strategies, and it is not clear which one will eventually win in
terms of performance. For now, though, we are limited by GCC, as it
mis-compiles two of the main contenders.

In this commit, the strategy we choose is the most conservative wrt.
what happened before. When a method should return a `long`, it
returns a `$Long` box with two fields `lo` and `hi`. This coincides
with the boxed representation of `long`s as `jl.Long!`. At call
site, we immediately unbox the result into two local variables for
the `lo` and `hi` words. Some care is taken not to do that if the
only usage of the resulting `long` pair is to be boxed again (e.g.,
to be further returned from the outer method).

All the methods of `RuntimeLong` explicitly take "expanded" versions
of their parameters: `abs` takes two parameters of type `Int`; `add`
takes 4. Shifts take 3 parameters of type `Int`: the `lo`, the `hi`,
and the shift amount. The *result*, however, is a `Long`.

In order to allow them to *construct* `Long` results from their
words, we introduce a magic method `RuntimeLong.pack(lo, hi)`, whose
body is filled in by the `Desugarer` with a special
`Transient(PackLong(lo, hi))`. It cannot be the compiler, because
we cannot serialize transients. And it cannot wait for the emitter,
because the optimizer definitely wants to see the `PackLong`s to
unpack them. An alternative would be to introduce a new `BinaryOp`,
but I think that's worse because it bakes an implementation detail
of the emitter into the IR. The fact that we *can* even do this
commit is a testament to the current abstraction level of our IR.

These changes significantly speed up even the SHA512 benchmark,
even though it performs most computations on stack already anyway
(the improvements must come from the arrays, in this case).

For the optimizer, this makes things a lot simpler. Instead of
having special-cases for `RuntimeLong` everywhere, we basically
have a unique `withSplitLong` method to deal with them. That
method can split one `PreTransform` of type `long` into two
`PreTransform`s of type `int`, so that they can be given to the
inlined methods of `RuntimeLong`. We introduce a new
`LongPairReplacement` for `LocalDef`s that aggregate a pair of
`(lo, hi)` `LocalDef`s (typically the result of a split
`PackLong`).

As a nice bonus, the IR checker now passes with `RuntimeLong`
after the optimizer.