Fix #5180: Fix race condition in class loading cycle detection
The key insight here is that we need an intermediate loading state
in which the infos are loaded but the class' parent chain is not
linked yet.

We must wait for all the infos of an inheritance cycle to be loaded to
detect the cycle reliably. However, we must not wait for the linking
to be completed, otherwise we end up with a future-cycle.