Fix #5208: Introduce raw floating point bit manipulation.
We repurpose the previous opcodes to mean the raw variants. They
do not provide any guarantee for the NaN bit patterns (other than
being one of the NaN patterns, obviously).

We now implement the canonicalizing variants in user-space on top
of the raw variants.

On Wasm, we had to do that anyway, except the "user-space" was in
the function emitter. On JavaScript, despite the spec "suggesting"
that NaN's be canonicalized, real-world engines do not canonicalize
all NaN's. Doing it in user-space is the only reliable way to
guarantee our spec.

---

We replace usages of the canonicalizing variants by raw variants in
the low-level javalib methods where we can. Either they are in code
paths where NaN's have been excluded, or where it does not actually
matter what bit pattern we receive for NaN's. This allows not to
regress on performance and code size for some low-level methods.
Unfortunately, `Double.hashCode` does require one more branch, but
that is inevitable to guarantee the correct semantics.

`ByteBuffer` methods `putFloat` and `putDouble` do not specify the
bit patterns of `NaN`s, and experimentally, I was able to observe
non-canonical bit patterns on the JVM. So in that case, we also use
the raw variants, which is consistent with using the `DataView`
methods in `TypedArray`-backed byte buffers.