Optimize CopyOnWriteArrayList to use scala.Array on Wasm
This commit optimizes `CopyOnWriteArrayList` to use a platform-specific
internal array (scala.Array on Wasm, and js.Array on JS),
similar to what was done for `ArrayList`.

On WebAssembly, it now uses a `scala.Array[AnyRef]` for its internal
storage. This improves performance by avoiding JS interop overhead for
array operations. On JavaScript, it continues to use `js.Array`.
InnerArrayImpl abstracts away the platform-specific details from the main class.

To avoid keeping track the length of array in the main class on Wasm,
we store the length of the array at the index 0 of the array.