Explicitly drop support for sbt < 1.9.0.
The new Maven Central Portal does not allow publishing artifacts under
the legacy sbt plugin path pattern. sbt < 1.9.0 will not be able to
resolve the new path patterns, so support for it is broken anyway.

We explicitly drop support for it internally, instead of pretending
that it will still work.