Wasm-friendly implementation of varargs.
For JavaScript, using Scala arrays wrapped in `WrappedArray` (2.12)
or `ArraySeq` (2.13) is not ideal for performance. Instead, since
about forever, we have re-emitted them as JS arrays wrapped in our
own `js.WrappedArray`/`WrappedVarArgs`. This, however, has poor
performance on Wasm.

We now call generic methods to choose the best implementation of
a varargs Seq, depending on the platform. They take Scala arrays,
and "convert" them to JS arrays if necessary. The conversions are
intrinsified so that they are zero-cost.

For this to work, we add an `InlineArrayReplacement` in the
optimizer, similar to the existing `InlineJSArrayReplacement`.