Wasm: Store the symbols of private JS fields on the JS side.
They always have to be manipulated on the JS side: at declaration
time in the custom JS helper that creates the JS class; and at
usage site for selection. Therefore, it was wasteful to store them
as Wasm globals, only to pass them to JS helpers every time.

We now declare them as JS variables, and we directly use those JS
variables from the JS class. For selection, we define a pair of
getter/setter helpers for each private JS field. This should be
more efficient than passing symbols from Wasm to a unique pair of
`jsSelect`/`jsSelectSet` helpers.