Use `js.Date.now()` in `System.{currentTimeMillis,nanoTime}`.
`js.Date.now()` was already present in ECMAScript 5.1. We were
using it in the constructor of `ju.Date`, but not in the two
methods of `System`.

In this commit, we switch to `js.Date.now()` in `System`. We also
change the constructor of `ju.Date` to delegate to
`currentTimeMillis`, in order to have a single source of truth.