Faster System.nanoTime(), both in JS and Wasm.
Instead of storing a lambda returning the high precision time,
we store the global object whose `now()` method returns the high
precision time. This is possible because the only two alternatives,
`performance.now()` and `Date.now()`, happen to share their method
name and the scale of their results.

This implementation removes a JS function call dispatch, which
makes the call faster both in JS and in Wasm.