Wasm: Opt: Fuse casts into custom JS helpers.
When the `expr` of a cast (including unchecked `asInstanceOf`s) is
a node implemented with a custom JS helper, we now fuse the
unboxing/cast operation into the JS-to-Wasm boundary, when possible.

This removes one round-trip to JS. For example, a call to
`js.Math.cos(x)` now results in a single Wasm-to-JS call, where the
JS helper has type `[f64] -> [f64]` in Wasm.