Fix #5246: Handle record trees that end up in statement position.
Usually, a record tree must always be used in a context that expects
the same record type. However, they can also end up in statement
position. Unfortunately, that means they can be put in a
`PreTransTree`. For safety reasons, `PreTransTree` aggressively
throws when given a tree with a `RecordType`, and that blows up in
this situation.

We could get rid of the safety net in `PreTransTree`, but that would
be a shame. Instead, when discarding the side effects of a tree, we
must make sure never to leave a tree with a `RecordType`. We do this
with a last resort of embedding it in a `Labeled` block with type
`void`. Since `Labeled` blocks cannot be removed, that is safe.

Once we're past the optimizer, the emitters were also missing some
handling of records in statement position:

* The JS emitter lacked a code path for returning a record value
  from a `void` function.
* The Wasm emitter was assuming that dropping a record always needs
  exactly one `drop`.