Refactoring: Introduce a typeclass for int/float/bits manipulation.
We now have a number of algorithms that need to be duplicated for
`Int`/`Long` or `Float`/`Double`. The new typeclass `IntFloatBits`
allows to generically write those algorithms only once, without the
cost of the abstraction.

This will make those algorithms more maintainable, as we don't have
to keep the copies in sync anymore.

We try not to change the underlying algorithms. There are two
exceptions where we reorder a few tests to have a single call to a
"configuration" anonymous function. That allows them to be fully
inlined. If we don't do that, the result of the anonymous functions
have to be boxed and unboxed.