Wasm: When enabled, use custom descriptors for the vtables.
Instead of storing the vtable as a user-space field, we fuse it
into the *descriptor* of our structs. This removes one word from
every allocated instance, reducing memory usage.

Described structs and descriptors must have a 1-1 correspondance
within a hierarchy. For regular Scala classes, this was already
the case. However we have to add:

* a fake `typeDataDescribed` struct to act as the type described
  by `typeData` (and it is therefore a supertype of `jl.Object`),
* distinct vtable types for the array classes.

The latter causes the most disruption across the emitter.
Everything else is fairly straightforward.

This does not use the JS interop aspects of the proposal yet.