Remove code paths for division by powers of 2 in RuntimeLong.
When the divisor is a *constant* power of 2, the optimizer now
rewrites the code ahead of time, so the run-time optimization is
not needed anymore.

When it is not constant, the chances that it happens to be a power
of 2 are extremely small. It is not worth having code paths for
them, which are slowing down the more common cases.