Remove the paths for small numerators in RuntimeLong divisions.
These were always a bit artificial. The scale of numerators is not
particularly stable, which means those fast paths were
unpredictable. They were worth it before because our slow, general
path was catastrophic.

Now that our general path is not so bad anymore, checking for the
(unpredictable) fast paths is not worth it anymore. It is better to
get as fast as possible to the general case.

Note that the general case has various code paths with different
performance characteristics. However, those depend on the scale of
the divisor, which is arguably more predictable. Anyway, those
paths are *necessary* for correctness, as opposed to being mere
optimizations.

Since we have fewer code, we can afford to directly inline the
unsigned operations into the signed ones. This way, there is only
a single function call for any of the 4 division operations.