Remove obsolete optimizations on long comparisons.
These optimizations haven't trigerred for `RuntimeLong`-based longs
since we started inlining them. And for Wasm, they are not that
important.

The only meaningful case was when we used `bigint`s for longs. But
even then, these optimizations had been over-fitted to the
initialization code of `Range`s. Now that `Range` doesn't use longs
at all, even that motivation is gone.

Since they are fairly specific and intricate, at this point I think
it is best to remove them.