Generic IntComparison in the optimizer.
Previously, we had a lot of duplicated (and not so duplicated)
code to handle integer comparisons on int/long and signed/unsigned.

We introduce `IntComparison` as a generic representation of any of
the 20 integer comparison operators. This allows to write the
folding logic only once, and actually cover all possible cases.