New encoding for the Unicode database.
We introduce a new encoding for the Unicode database that is
more efficient in terms of code size, memory consumption and
performance.

In addition, it more faithfully encodes some of the Unicode
properties. In particular, `isLowerCase` and `isUpperCase` now
take into account code points with the `Other_LowerCase` and
`Other_UpperCase` contributory properties.

Where previously we only recorded the *type* of code point ranges,
we now also store the value of 5 property flags. For the first
160 code points, which are stored in a direct array, we store the
value of 7 properties.

Finally, we support an encoding of ranges with "alternating types",
where odd code points have one type and even code points have
another type. This is useful for many ranges of the Unicode space.
Most notably, there are many ranges with alternating LowerCase (Ll)
and UpperCase (Lu) types.

We store each range as a clever bit field:

    rarePropFlags (x) ++ firstCP (21) ++ frequentPropFlags (5 - x)
      ++ alternatingTypes flag (1) ++ type (5)

which, remarkably, adds up to 32 bits. A single array therefore
stores everything, whereas previously we separately stored the
range start indices from the corresponding types.

The 5 properties that we store cannot be derived from the type
field, but are used in conjunction with the type field to implement
some of the `isXYZ` methods. A single, efficient method
`dataHasAnyFlag` can test for a set of types and a set of properties.
For example, `isAlphabetic` returns `true` when either the type is
any of the five major `XYZ_LETTER` types or `LETTER_NUMBER`, or when
any of the properties `Other_LowerCase`, `Other_UpperCase` or
`Other_Alphabetic` is on.

Two `isXYZ` methods, namely `isIdeographic` and `isMirrored`, only
rely on a boolean property, and not on the type. Since those
properties are not even very correlated to the ranges used for the
main data, they receive their independent arrays of ranges.

Since the logic is bit-hackery-heavy, we also generate exhaustive
test data with straightforward logic.